requirements.txt
absl-py>=1.0.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2023.5.7
chardet==4.0.0
click==8.0.3
numpy==1.23.5
colorama==0.4.6
Flask==2.2.2
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib>=0.5
google-pasta==0.2.0
grpcio>=1.48.2
h5py==3.7.0
idna==3.4
importlib-metadata==4.13.0
itsdangerous==2.1.1
Jinja2==3.1.2
joblib==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.0
opt-einsum==3.3.0
pickle-mixin==1.0.2
Pillow==9.2.0
protobuf>=3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy>=1.9.0
six==1.16.0
tensorboard==2.12.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
